feat: Add functions to check exit status and order status for CookingPersonel
The code changes add new functions to the `CookingPersonel` class to check the exit status and order status. These functions will be used to determine if the personnel should exit their respective threads and if there are any active or pending orders. This improves the control and management of the cooking personnel and their associated orders. Previous design was more complex and buggy.